AddressBook.java: add support for finding names with case insensitive input
The current program requires users to search for people by inputting their names in the exact casing that was provided.

This may cause some inconvenience as users may not be aware of the casing that was used to store the names.

This commit modifies the find feature to return all names that have the same English characters but possibly with different casings.